+++ bicleaner/bicleaner_train.py
@@ -6,7 +6,9 @@ from sklearn import neighbors
 from sklearn import svm
 from sklearn.ensemble import AdaBoostClassifier
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.externals import joblib
+#from sklearn.externals import joblib
+import joblib
+
 from tempfile import TemporaryFile, NamedTemporaryFile
 from timeit import default_timer
 
@@ -18,19 +20,18 @@ import os
 import random
 import sklearn
 import sys
+import json
 from toolwrapper import ToolWrapper
 from mosestokenizer import MosesTokenizer
 
-#Allows to load modules while inside or outside the package 
-
 #Allows to load modules while inside or outside the package  
 try:
-    from .features import feature_extract, FEATURES_VERSION
+    from .features import feature_extract, FEATURES_VERSION, Features
     from .prob_dict import ProbabilisticDictionary
     from .util import no_escaping, check_positive, check_positive_or_zero, logging_setup
     from .training import shuffle,precision_recall, repr_right, write_metadata, train_fluency_filter
 except (SystemError, ImportError):
-    from features import feature_extract, FEATURES_VERSION
+    from features import feature_extract, FEATURES_VERSION, Features
     from prob_dict import ProbabilisticDictionary
     from util import no_escaping, check_positive, check_positive_or_zero, logging_setup
     from training import shuffle,precision_recall, repr_right, write_metadata, train_fluency_filter 
@@ -41,9 +42,16 @@ __author__ = "Sergio Ortiz-Rojas"
 __version__ = "Version 0.1 # December 2017 # Initial version # Sergio Ortiz-Rojas"
 __version__ = "Version 0.2 # 09/01/2018 # Adding argument for injecting wrong examples from a file # Jorge Ferrández-Tordera"
 __version__ = "Version 0.3 # 18/01/2019 # Integrated training of LM and refactor to avoid code duplicity # Víctor M. Sánchez-Cartagena"
+__version__ = "Version 0.13 # 30/10/2019 # Features version 3  # Marta Bañón"
+
+
+logging_level = 0
     
 # Argument parsing
 def initialization():
+
+    global logging_level
+    
     parser = argparse.ArgumentParser(prog=os.path.basename(sys.argv[0]), formatter_class=argparse.ArgumentDefaultsHelpFormatter, description=__doc__)
 
     parser.add_argument('input',  nargs='?', type=argparse.FileType('r'), default=sys.stdin,  help="Tab-separated bilingual input file")
@@ -73,6 +81,7 @@ def initialization():
     groupO.add_argument('-p', '--processes', type=check_positive, default=max(1, cpu_count()-1), help="Number of process to use")
     groupO.add_argument('--wrong_examples_file', type=argparse.FileType('r'), default=None, help="File with wrong examples extracted to replace the synthetic examples from method used by default")
     groupO.add_argument('--features_version', type=check_positive, default=FEATURES_VERSION , help="Version of the features")
+    groupO.add_argument('--disable_lang_ident', default=False, action='store_true', help="Don't apply features that use language detecting")
 
     #For LM filtering
     groupO.add_argument('--noisy_examples_file_sl', type=str, help="File with noisy text in the SL. These are used to estimate the perplexity of noisy text.")
@@ -94,6 +103,15 @@ def initialization():
     args = parser.parse_args()
     # Logging
     logging_setup(args)
+    
+    logging_level = logging.getLogger().level    
+    
+    if logging_level <= logging.WARNING and logging_level != logging.DEBUG:
+        #Getting rid of INFO messages when Moses processes start
+        logging.getLogger("MosesTokenizer").setLevel(logging.WARNING)
+        logging.getLogger("MosesSentenceSplitter").setLevel(logging.WARNING)
+        logging.getLogger("MosesPunctuationNormalizer").setLevel(logging.WARNING)
+    
     return args
 
 # Training function: receives two file descriptors, input and test, and a
@@ -144,6 +162,14 @@ def train_classifier(input_features, test_features, classifier_type, classifier_
 
     clf.fit(dataset['data'], dataset['target'])
 
+    # Log sorted feature importances with their names
+    if classifier_type in ('random_forest', 'adaboost'):
+        feat_names = Features.cols + Features.optional
+        feat_dict = dict(zip(feat_names, clf.feature_importances_))
+        sorted_feat = {k: v for k, v in sorted(feat_dict.items(), key=lambda item: item[1])}
+    else:
+        sorted_feat = None
+
     joblib.dump(clf, classifier_output)
 
     feats = []
@@ -170,7 +196,7 @@ def train_classifier(input_features, test_features, classifier_type, classifier_
     hgood  = np.histogram(good,  bins = np.arange(0, 1.1, 0.1))
     hwrong = np.histogram(wrong, bins = np.arange(0, 1.1, 0.1))
 
-    return hgood[0].tolist(), hwrong[0].tolist()
+    return hgood[0].tolist(), hwrong[0].tolist(), sorted_feat
 
 
 
@@ -291,8 +317,8 @@ def map_process(input, block_size, jobs_queue, label, first_block=0):
 # Main loop of the program
 def perform_training(args):
     time_start = default_timer()
-    logging.info("Starting process")
-    logging.info("Running {0} workers at {1} rows per block".format(args.processes, args.block_size))
+    logging.debug("Starting process")
+    logging.debug("Running {0} workers at {1} rows per block".format(args.processes, args.block_size))
 
     process_count = max(1, args.processes)
     maxsize = 1000 * process_count
@@ -302,10 +328,17 @@ def perform_training(args):
 
     #Read input to a named temporary file
     #We may need to read it multiple times and that would be problematic if it is sys.stdin
+
+    count_input_lines = 0
     input = NamedTemporaryFile(mode="w",delete=False)
     for line in args.input:
         input.write(line)
+        count_input_lines = count_input_lines +1
     input.close()
+
+    if count_input_lines < 10000:
+        logging.error("Training corpus must be at least 10K sentences long (was {}).".format(count_input_lines))
+        sys.exit(1)
     
     stats=None
     with open(input.name) as input_f:
@@ -393,12 +426,14 @@ def perform_training(args):
         
         features_train.seek(0)
         features_test.seek(0)
-        hgood, hwrong = train_classifier(features_train, features_test, args.classifier_type, args.classifier)
+        hgood, hwrong, feat_importances = train_classifier(features_train, features_test, args.classifier_type, args.classifier)
         features_train.close()
         features_test.close()
 
     logging.info("End training")
 
+    if feat_importances is not None:
+        logging.debug('Feature importances: ' + json.dumps(feat_importances, indent=4))
     write_metadata(args, length_ratio, hgood, hwrong, stats)
     args.metadata.close()
 
@@ -417,4 +452,5 @@ def main(args):
 
 if __name__ == '__main__':
     args = initialization()
+    print(args.classifier_type)
     main(args)