+++ CHANGELOG.md
@@ -1,5 +1,24 @@
+Bicleaner 0.13:
+* Bicleaner hardrules changes:
+  * Rule change: Relaxed c_minimal_length to accept 3-word sentences	
+  * New feature: LM filtering (moved from Bicleaner Classify)
+  * New parameter: `--disable_lm_filter`, `--metadata` and `--lm_threshold`, to support LM filtering
+* Bicleaner training changes: 
+  * New parameter: Features relying on language identification can be disabled with flag `--disable_lang_ident` (this will be outputed in the .yaml file and used by Bicleaner clasifier)
+  * New feature: Debug mode now gives information on random forest feature importances
+  * Parameter change: --noisy_examples_file_sl and --noisy_examples_file_tl are now optional
+  * Parameter change: input now must be more than 10K sentences long
+  * Removed INFO messages when processes starting/ending (except when debugging)
+* Bicleaner classifier changes:
+  * `--disable_lang_ident` flag is now read from the .yaml file
+  * Removed feature: LM filtering (moved to Bicleaner Hardrules)
+  * New parameter: `--disable_lm_filter`
+  * Removed parameter: `--keep_lm_result`
+* Other:
+  * Updated requirements
+  
+  
 Bicleaner 0.12:
-
 * Bicleaner hardrules changes:
   * New rule: c_identical_wo_punct to reject sentences only different in punctuation (and it's case insensitive)
   * New rule:  Sentences containing "Re:" are rejected
@@ -9,7 +28,7 @@ Bicleaner 0.12:
   * Rule change: Breadcrumbs2 now includes character "·" in the rejected characters
   * Rule change: c_length now compares byte length ratio (will avoid rejecting valid sentences due to length ratio when comparing languages with different alphabets)
   * Changed behaviour for `--annotated_output` argument in hardrules. See README.md for more information.
-  * New parameter: '--disable_lang_ident` flag to avoid applying rules that need to identify the language
+  * New parameter: `--disable_lang_ident` flag to avoid applying rules that need to identify the language
 * Bicleaner classify changes:  
   * Now using only 3 decimal places for Bicleaner score and LM score
   * Removed INFO messages when processes starting/ending (except when debugging)